move util file
